[dynamo][source] Remove index_is_slice from GetItemSource
Slicing on the data structures that go through GetItemSource always
result in temporary objects. So, there is never a case where we need a
source for a sliced data structure. I think it was a bug at one point of
time, and we incorrectly provided a source for sliced data structures.

ghstack-source-id: 30938a8f258edc3d7d2876dd5bb9e71eda1d19b4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/146887